+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.32] - 2025-12-27
 
+- ✨ Add CSV tools (CSV <-> Markdown Table)
+- ✨ Add Data Structure tools (Env to JSON)
+- ✨ Add JSON tools (Flatten, Unflatten)
+- ✨ Add XML tools (XML to JSON)
+- ✨ Add URL tools (URL Params to JSON)
 - ✨ Add Text Cleanup tools (Remove Empty Lines, Remove Line Numbers, Join Lines, Split Lines, Trim Lines)
 - ✨ Add Math Statistics tools (Sum, Average, Min, Max)
 - ✨ Add Random Generators (Password, IPv4, IPv6)